Replace dependency on `dirs` with `home`
The `dirs` crate recently started depending on the `options-ext` crate
which uses copyleft license (MPL). This (unnecessary) dependency causes
licensing issues for various users by possibly poisoning the dependency
tree of their projects[1].

This change replaces the `dirs` crate with `home`. The `home` crate is
maintained by the cargo team and offers the same functionality.

As a bonus, this change also results in a slightly smaller dependency
tree.

[1]:
- https://github.com/artichoke/artichoke/pull/2564
- https://github.com/pyrossh/rust-embed/issues/231
- https://github.com/juhaku/utoipa/issues/834
- https://github.com/harryfei/which-rs/pull/78